Move the engine to its own file
It makes easy to understand the project structure.
